Move CsrfToken.js to Common.js
And change it to the non-default export.

This sets the stage for other common functions I want to add.
