+++ frontend/src/controllers/Common.js
@@ -1,5 +1,5 @@
 // Retrieve the CSRF token from the meta tag in the DOM.
-export default function getCsrfToken() {
+export function getCsrfToken() {
   return document
     .querySelector("meta[name='csrf-token']")
     .getAttribute('content');
